tp/requirements.txt
coverage
selenium
python-dotenv